refactor: Move lib packages to internal packages
