+++ cmd/server/start.go
@@ -7,8 +7,8 @@ import (
 	"os/user"
 
 	"github.com/glass-cms/glasscms/database"
+	ctx "github.com/glass-cms/glasscms/internal/context"
 	"github.com/glass-cms/glasscms/item"
-	ctx "github.com/glass-cms/glasscms/lib/context"
 	"github.com/glass-cms/glasscms/server"
 	"github.com/glass-cms/glasscms/server/handler"
 	v1 "github.com/glass-cms/glasscms/server/handler/v1"
