+++ internal/log/noop.go
