+++ internal/middleware/content_type.go
@@ -4,7 +4,7 @@ import (
 	"net/http"
 	"slices"
 
-	"github.com/glass-cms/glasscms/lib/mediatype"
+	"github.com/glass-cms/glasscms/internal/mediatype"
 )
 
 // ContentType generates a handler that writes a 415 Unsupported Media Type header
