+++ internal/middleware/content_type_test.go
@@ -5,8 +5,8 @@ import (
 	"net/http/httptest"
 	"testing"
 
-	"github.com/glass-cms/glasscms/lib/mediatype"
-	"github.com/glass-cms/glasscms/lib/middleware"
+	"github.com/glass-cms/glasscms/internal/mediatype"
+	"github.com/glass-cms/glasscms/internal/middleware"
 )
 
 func Test_ContentType(t *testing.T) {
