+++ internal/resource/resource.go
