+++ server/handler/v1/errors_test.go
@@ -5,7 +5,7 @@ import (
 	"testing"
 
 	v1 "github.com/glass-cms/glasscms/api/v1"
-	"github.com/glass-cms/glasscms/lib/resource"
+	"github.com/glass-cms/glasscms/internal/resource"
 	v1_handler "github.com/glass-cms/glasscms/server/handler/v1"
 	"github.com/stretchr/testify/require"
 )
